Ajustado botões para finalizar, excluir e estornar venda; Implementado funcionalidade para adicionar unidades a produto;

+++ src/app/vendas/venda-avulsa/venda-avulsa.component.ts
@@ -13,6 +13,8 @@ import { environment } from './../../../environments/environment';
 import { ClienteService } from 'src/app/clientes/cliente.service';
 
 import { ToastyService } from 'ng2-toasty';
+import { Produto } from 'src/app/model/Produto';
+import { VendaProduto } from 'src/app/model/VendaProduto';
 
 @Component({
   selector: 'app-venda-avulsa',
@@ -20,6 +22,9 @@ import { ToastyService } from 'ng2-toasty';
   styleUrls: ['./venda-avulsa.component.css']
 })
 export class VendaAvulsaComponent implements OnInit {
+
+  exclusaoProduto = 'exclusaoProduto';
+  produtoModal;
 
   titulo: any;
   produtosPesquisa = [];
@@ -32,8 +37,10 @@ export class VendaAvulsaComponent implements OnInit {
 
   @ViewChild('inputProduto', {read: ElementRef}) inputProduto: any;
   @ViewChild('divLiveSearch', {read: ElementRef}) divLiveSearch: any;
+  @ViewChild('botaoFinalizar', {read: ElementRef}) botaoFinalizar: any;
 
   formulario: FormGroup;
+  formularioTemp = [];
   formBuilder = new FormBuilder();
 
   tempo: any;
@@ -67,7 +74,42 @@ export class VendaAvulsaComponent implements OnInit {
     } else {
       this.titulo = '';
     }
+    this.botaoFinalizar.nativeElement.disabled = true;
+    this.onChange();
+  }
 
+  onChange() {
+    const array = ['clienteId', 'desconto', 'observacao'];
+    this.formulario.valueChanges.subscribe(val => {
+      for (const prop in this.formulario.value) {
+        if (prop) {
+          if (array.indexOf(prop) >= 0 &&
+          this.formulario.value[prop] !== this.formularioTemp[prop]) {
+            this.botaoFinalizar.nativeElement.disabled = true;
+            break;
+          } else {
+            this.botaoFinalizar.nativeElement.disabled = false;
+          }
+          if (prop === 'produtos') {
+            for (const {item, index} of this.formulario.value[prop].map((item, index) => ({item, index})) ) {
+              if (this.formularioTemp['produtos'].length > 0) {
+                if (Number(item.produto.quantidade) !== Number(this.formularioTemp['produtos'][index].quantidade)) {
+                  this.botaoFinalizar.nativeElement.disabled = true;
+                  break;
+                } else {
+                  this.botaoFinalizar.nativeElement.disabled = false;
+                }
+              } else {
+                this.botaoFinalizar.nativeElement.disabled = true;
+              }
+            }
+          }
+          if (this.botaoFinalizar.nativeElement.disabled) {
+            break;
+          }
+        }
+      }
+    });
   }
 
   carregarVenda(id: number) {
@@ -79,6 +121,7 @@ export class VendaAvulsaComponent implements OnInit {
           this.titulo = '#' + id;
         }
         this.preencherProdutosFormulario(response.produtos);
+        this.ativarBotaoFinalizar();
       })
       .catch(erro => this.errorHadler.handle(erro));
   }
@@ -94,8 +137,22 @@ export class VendaAvulsaComponent implements OnInit {
         }
         this.toasty.success('Venda salva!');
         this.preencherProdutosFormulario(response.produtos);
+        this.ativarBotaoFinalizar();
       })
       .catch(error => this.errorHadler.handle(error));
+  }
+
+  ativarBotaoFinalizar() {
+    this.formularioTemp = this.formulario.value;
+    this.formularioTemp['produtos'] = [];
+    for (const item of this.formulario.get('produtos').value) {
+      const vp = new VendaProduto();
+      vp.quantidade = item.quantidade;
+      vp.id = item.id;
+      vp.produto = item.produto;
+      this.formularioTemp['produtos'].push(vp);
+    }
+    this.botaoFinalizar.nativeElement.disabled = false;
   }
 
   finalizarVenda() {
@@ -134,7 +191,11 @@ export class VendaAvulsaComponent implements OnInit {
   }
 
   inputValorProdutoAlterado(event, item) {
-    item.quantidade = event.target.value;
+    let valor = event.target.value;
+    if (valor > (item.produto.estoque - item.produto.reserva)) {
+      valor = item.produto.estoque - item.produto.reserva;
+    }
+    item.quantidade = valor;
 
     this.atualizarVenda();
   }
@@ -143,7 +204,12 @@ export class VendaAvulsaComponent implements OnInit {
     return this.formulario.get('valor').value - this.formulario.get('desconto').value;
   }
 
-  excluirItem(item: any) {
+  retirarItemDaLista(idModal, produto) {
+    this.produtoModal = produto;
+    this.openModal(idModal, this.produtoModal);
+  }
+
+  excluirItem(item: any, idModal) {
     const p1 = this.formulario.get('produtos').value;
 
     const novaLista = p1.filter(p => {
@@ -156,9 +222,15 @@ export class VendaAvulsaComponent implements OnInit {
     novaLista.forEach(nl => p2.push(nl));
 
     const idFormulario = this.formulario.get('id').value;
-    if (idFormulario) {
-      this.vendasService.removerProduto(idFormulario, item.produto.id);
+    if (idFormulario && item.id) {
+      this.vendasService.removerProduto(idFormulario, item.produto.id)
+        .then(response => {
+          this.formulario.patchValue(response);
+          this.preencherProdutosFormulario(response.produtos);
+          this.ativarBotaoFinalizar();
+        });
     }
+    this.closeModal(idModal);
   }
 
   selecionaProduto(event) {
@@ -253,6 +325,9 @@ export class VendaAvulsaComponent implements OnInit {
       observacao: [''],
       status: ['']
     });
+
+    this.formularioTemp = this.formulario.value;
+    //console.log('testmp', this.formularioTemp);
   }
 
   removeDivChildren(div: any) {
@@ -276,22 +351,27 @@ export class VendaAvulsaComponent implements OnInit {
       const liHead = this.render2.createElement('div');
       this.render2.setAttribute(liHead, 'class', 'row');
       const htNome = this.render2.createText('Nome');
-      const htEstoque = this.render2.createText('Unidades disponíveis');
+      const htEstoque = this.render2.createText('Estoque');
+      const htReserva = this.render2.createText('Reserva');
       const htValor = this.render2.createText('Valor R$');
 
       const hsNome = this.render2.createElement('span');
-      this.render2.setAttribute(hsNome, 'class', 'col-md-7');
+      this.render2.setAttribute(hsNome, 'class', 'col-md-6');
       const hsEstoque = this.render2.createElement('span');
-      this.render2.setAttribute(hsEstoque, 'class', 'col-md-3 text-center');
+      this.render2.setAttribute(hsEstoque, 'class', 'col-md-2 text-center');
+      const hsReserva = this.render2.createElement('span');
+      this.render2.setAttribute(hsReserva, 'class', 'col-md-2 text-center');
       const hsValor = this.render2.createElement('span');
       this.render2.setAttribute(hsValor, 'class', 'col-md-2 text-center');
 
       this.render2.appendChild(hsNome, htNome);
       this.render2.appendChild(hsEstoque, htEstoque);
+      this.render2.appendChild(hsReserva, htReserva);
       this.render2.appendChild(hsValor, htValor);
 
       this.render2.appendChild(liHead, hsNome);
       this.render2.appendChild(liHead, hsEstoque);
+      this.render2.appendChild(liHead, hsReserva);
       this.render2.appendChild(liHead, hsValor);
       this.render2.appendChild(ulhead, liHead);
 
@@ -310,21 +390,26 @@ export class VendaAvulsaComponent implements OnInit {
         this.render2.setAttribute(li, 'class', 'body-live-list row');
         const nome = this.render2.createText(p.nome);
         const estoque = this.render2.createText(p.estoque);
+        const reserva = this.render2.createText(p.reserva);
         const valor = this.render2.createText(p.valor);
 
         const spanNome = this.render2.createElement('span');
-        this.render2.setAttribute(spanNome, 'class', 'col-md-7');
+        this.render2.setAttribute(spanNome, 'class', 'col-md-6');
         const spanEstoque = this.render2.createElement('span');
-        this.render2.setAttribute(spanEstoque, 'class', 'col-md-3 text-center');
+        this.render2.setAttribute(spanEstoque, 'class', 'col-md-2 text-center acert-padding');
+        const spanReserva = this.render2.createElement('span');
+        this.render2.setAttribute(spanReserva, 'class', 'col-md-2 text-center acert-padding');
         const spanValor = this.render2.createElement('span');
-        this.render2.setAttribute(spanValor, 'class', 'col-md-2 text-center');
+        this.render2.setAttribute(spanValor, 'class', 'col-md-2 text-center acert-padding');
 
         this.render2.appendChild(spanNome, nome);
         this.render2.appendChild(spanEstoque, estoque);
+        this.render2.appendChild(spanReserva, reserva);
         this.render2.appendChild(spanValor, valor);
 
         this.render2.appendChild(li, spanNome);
         this.render2.appendChild(li, spanEstoque);
+        this.render2.appendChild(li, spanReserva);
         this.render2.appendChild(li, spanValor);
         this.render2.appendChild(ulbody, li);
       }
@@ -336,7 +421,7 @@ export class VendaAvulsaComponent implements OnInit {
 
   }
 
-  openModal(id: string, botaoExcluir: any, tipo: any) {
+  openModal(id: string, produto: any) {
     this.modalService.open(id);
     this.inputProduto.nativeElement.focus();
   }
